refactor: restructure repository for microservices
