+++ .gitignore
@@ -1,4 +1,3 @@
-src/__pycache__
 __pycache__
 cam1.mp4
 cam2.mp4
@@ -6,9 +5,7 @@ cam2.mp4
 pose_landmarker_full.task
 yolov8n-pose.pt
 yolov8s-pose.pt
-__pycache__
-services/analytics_service/__pycache__
-services/camera_service/__pycache__
-services/user_service/__pycache__
-src/__pycache__
-src/influx_conf/__pycache__
+microservices/analytics_service/__pycache__
+microservices/camera_service/__pycache__
+microservices/user_service/__pycache__
+common/influx_conf/__pycache__
