+++ common/openapi: 3.0.yml
